Refactor Contact/Deal Properties to share base class.
