+++ lib/hubspot-ruby.rb
@@ -6,6 +6,7 @@ require 'hubspot/company'
 require 'hubspot/config'
 require 'hubspot/connection'
 require 'hubspot/contact'
+require 'hubspot/properties'
 require 'hubspot/contact_properties'
 require 'hubspot/contact_list'
 require 'hubspot/form'
